+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Added the ability to paginate the relationships of a profile ([#467](https://github.com/desmos-labs/desmos/issues/467))
 - Added the ability to paginate user blocks ([#495](https://github.com/desmos-labs/desmos/issues/495))
 - Added the ability to paginate incoming DTag transfer requests ([#519](https://github.com/desmos-labs/desmos/pull/519))
+- Renamed the following messages
+   - `MsgCancelDTagTransfer` -> `MsgCancelDTagTransferRequest` 
+   - `MsgAcceptDTagTransfer` -> `MsgAcceptDTagTransferRequest`
+   - `MsgRefuseDTagTransfer` -> `MsgRefuseDTagTransferRequest`
 - Added the possibility to connect external chain accounts to a Desmos profile ([#192](https://github.com/desmos-labs/desmos/issues/192))  
 - Added the possibility to verify a profile with an external application ([#472](https://github.com/desmos-labs/desmos/issues/472))
 - Added the ability to edit whether users can comment on a post or not ([#446](https://github.com/desmos-labs/desmos/issues/446))